Move simplex noise to its own file Fix a warning avoid void* -> int32 on x64
