+++ src/tools/quemap/simplex.h
@@ -1,0 +1,45 @@
+/*
+ * Copyright(c) 1997-2001 id Software, Inc.
+ * Copyright(c) 2002 The Quakeforge Project.
+ * Copyright(c) 2006 Quetoo.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ */
+
+#pragma once
+
+/**
+ * 3D Perlin simplex noise
+ *
+ * @param[in] x float coordinate
+ * @param[in] y float coordinate
+ * @param[in] z float coordinate
+ *
+ * @return Noise value in the range[-1; 1], value of 0 on all integer coordinates.
+ */
+float SimplexNoise(float x, float y, float z);
+
+/**
+ * Fractal/Fractional Brownian Motion (fBm) summation of 3D Perlin Simplex noise
+ *
+ * @param[in] octaves   number of fraction of noise to sum
+ * @param[in] x         x float coordinate
+ * @param[in] y         y float coordinate
+ * @param[in] z         z float coordinate
+ *
+ * @return Noise value in the range[-1; 1], value of 0 on all integer coordinates.
+ */
+float SimplexNoiseFBM(size_t octaves, float frequency, float amplitude, float lacunarity, float persistence, float x, float y, float z);
